+++ packages/transaction-pay-controller/src/utils/source-amounts.ts
@@ -1,4 +1,5 @@
 import { TransactionType } from '@metamask/transaction-controller';
+import type { TransactionMeta } from '@metamask/transaction-controller';
 import { createModuleLogger } from '@metamask/utils';
 import { BigNumber } from 'bignumber.js';
 
@@ -7,7 +8,6 @@ import type {
   TransactionPaymentToken,
 } from '..';
 import { TransactionPayStrategy } from '..';
-import type { TransactionMeta } from '../../../transaction-controller/src';
 import { ARBITRUM_USDC_ADDRESS, CHAIN_ID_ARBITRUM } from '../constants';
 import { projectLogger } from '../logger';
 import type {